alterado o tipo do contador para long int e o titulo do arquivo

+++ Algoritmos-ordenação/Ordenacao.h
@@ -6,19 +6,19 @@
 #include <time.h>
 
 void swap(int *fp , int *sp);
-int partition(int *arr,int begin,int end,int conta[]);
+int partition(int *arr, int begin, int end, long  int conta[]);
 
-void BubbleSort(int vetor[],int n,int conta[]);
-void SelectSort(int vetor[],int n,int []);
-void InsertSort(int vetor[],int n,int []);
-void MergeSort(int vet[],int inicio,int fim, int conta []);
-void merge(int vet[],int inicio,int m,int fim,int conta[]);
-void quickSort(int arr[],int begin,int end,int conta []);
+void BubbleSort(int vetor[], int n, long int conta[]);
+void SelectSort(int vetor[], int n, long  int[]);
+void InsertSort(int vetor[], int n, long int[]);
+void MergeSort(int vet[], int inicio, int fim, long  int conta[]);
+void merge(int vet[], int inicio, int m, int fim, long  int conta[]);
+void quickSort(int arr[], int begin, int end, long int conta[]);
 
 void Preenche(int arr[], int n, int op);
 void menu();
-double Escolhe(int arr[],int n,int op,int conta[]);
-void resultado(double cptime, int conta[], int ordenacao, int tipo, int length);
+double Escolhe(int arr[], int n, int op, long  int conta[]);
+void resultado(double cptime, long int conta[], int ordenacao, int tipo, int length);
 void teste(int ordem, int tipo, int n);
 
 #endif
